Aggiunta metodo e classe per calcoloBaricentro

+++ src/it/unibs/fp/sistemaStellare/Posizione.java
@@ -1,28 +1,28 @@
 package it.unibs.fp.sistemaStellare;
 
 public class Posizione {
-    private int x, y;
+    private float x, y;
 
     public Posizione () {}
 
-    public Posizione (int x, int y) {
+    public Posizione (float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
